# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EBOS API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login Manaul",
                "operationId": "loginManual",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AuthArray"
                },
                "responses": {
                    "200": {
                        "description": "Returns JWT"
                    }
                }
            }
        },
        "/api/v1/refreshtoken": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh Token",
                "operationId": "refreshToken",
                "responses": {
                    "200": {
                        "description": "Returns JWT"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "requestBodies": {
            "AuthArray": {
                "description": "Auth Paramerters",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "email": {
                                    "title": "Email",
                                    "description": "Email",
                                    "format": "string",
                                    "example": "matt@gmail.com"
                                },
                                "password": {
                                    "title": "Password",
                                    "description": "password",
                                    "format": "string",
                                    "example": "a123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}